+++ test/internal/db/schema.rb
@@ -50,6 +50,7 @@ ActiveRecord::Schema.define do
     t.text :info
     t.text :info2_ciphertext
     t.text :city_ciphertext
+    t.binary :ssn_ciphertext
   end
 
   create_table :posts do |t|